pyproject.toml
[project]
name = "self-initiated-research-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.15",
    "arxiv>=2.2.0",
    "groq>=0.31.0",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.74",
    "langchain-groq>=0.3.7",
    "langgraph>=0.6.4",
    "python-dotenv>=1.1.1",
    "tavily-python>=0.7.10",
]
